Add pre-sieving
The pattern of multiples of prime numbers repeats.  This is what allows
for wheel factorization, and likewise it allows for "pre-sieving," where
the pattern of multiples of the first few prime numbers is determined
ahead of time and then copied into the bitset during sieving later.
Like a wheel, this means that small primes (with the most multiples) do
not have to be considered during the sieve itself (as they are already
accounted for), significantly speeding up computations.

This idea is already employed in several other prime programs.

This adds an initial version of pre-sieving to yase.  It still requires
documentation and refinement, but the "plumbing" is in place.