Adjust multiple marking for slightly improved performance
This generates a slight performance increase both with GCC and Clang on
my machine.  I think this is because it moves similar operations farther
apart in the actual source code, guiding the compiler to produce code
that is a bit more friendly to instruction-level parallelism.
Admittedly, it's fragile, but it appears to work, and it does not change
the program from a logical perspective, so there you go.