Rename multiple to prime, keep track of index instead of wheel_elem
Each struct multiple really represents a sieving prime, so rename it to
struct prime.

Keep track of the index in the wheel table for each sieving prime, not
the pointer to the table entry.  This should better facilitate using a
very tight multiple-marking loop for small sieving primes in the future.