Introduce highly-optimized small prime sieve
Now, sieving primes such that prime->prime_adj < SMALL_THRESHOLD will be
sieved using a highly-optimized set of mod 30 multiple-marking loops
that essentially encode the wheel30 table into the actual machine
instructions.  This creates a very significant performance boost when
used for primes with enough multiples per segment to outweigh the
overhead.  See src/sieve.c for more information.

When finding sieving primes in src/seed.c, two lists are generated:
one of the primes to be sieved with the small sieve, and one of the
primes to be sieved with the large sieve.

This code is based closely on the unbelievably fast prime-sieving
program "primesieve," which employs a set of loops like this to mark
multiples of small primes.  Thank you, primesieve!