Tweak small sieve for performance
If we keep track of the actual byte in the sieve to mark, we can get a
healthy performance increase in the small sieve.  This is the way that
primesieve does it.  Sigh . . . how is it that primesieve is always
right?

This produces nicer-looking and more concise assembly, with the or
instructions to mark the multiples all in a row.  Beautiful.