Use 32-bit byte indices, 32-bit prime counter when sieving
The index of the byte to mark should never required 64 bits (as that
would mean a very long sieve bit array!), so locally use only 32 bits to
keep track of this value.  This produces a big performance boost and
32-bit systems.

Additionally, use a local counter in sieve_segment() and only update the
passed counter once.  On 32-bit systems this removes the overhead of
having to add repeatedly to a uint64_t.