Only allocate as many list head pointers as necessary
Memory usage is reduced to mainly the storage of sieving primes by only
allocating as many list head pointers as is necessary (i.e. calculating
the maximum number that will be needed ahead of time).

As part of these changes, the entire segment (of SEGMENT_BYTES bytes) is
always sieved, while only the relevant portion of the segment is
examined to count primes.  Additionally, no "end-checking" to determine
if a sieving prime's next segment is within the interval being sieved
is necessary.  (The changes to the list-head-pointer system mean that
now there is always a list available to place sieving primes in, even if
that list is for a segment outside of the interval being sieved.)  This
makes the code a little cleaner in these areas.