Allow mathematical expressions in command line
Using the facilities provided in the new src/expr.c, the maximum value
command line argument will be evaluated as a mathematical expression.
The evaluator supports addition, subtraction, multiplication, and
exponentiation, as well as "e" notation.

The evaluator consists of a lexer and what is essentially a
recursive-descent parser that evaluates the expression as it parses it.