Restructure sieving of the entire requested interval
src/interval.c now handles the sieving of an entire interval, including
breaking the interval into segments and calculating the beginning and
end bits and bytes.  In order to simplify things, prime counting is no
longer done in src/seed.c.  Instead, the main sieve is run for the
entire requested interval.  There is some duplication of work, but it
makes the process much simpler with little impact on performance.