Accept interval minimum value on command line
If two arguments are given, they are interpreted as the minimum and
maximum; if only one is given, the argument is interpeted as the maximum
and the minimum is assumed to be 0.

These changes incorporate the inclusion of a pi(x) table for x < 30.
One feature associated with this is that maximum values under 30 now
work just fine.