Process small sieving primes by next wheel_idx
Previously, when the multiples of small sieving primes were marked in a
segment, the sieving primes were processed in whatever order they were
added to the small prime list.  If instead we process small sieving
primes batched by next wheel_idx, then all of the jumps to case 0 in the
big 64-way switch in sieve.c will occur first, then all of the jumps to
case 1, then all of the jumps to case 2, and so on. As a result, the
CPU's branch predictor will predict the correct branches more
frequently, giving us a nice speed-up.

This idea, like many of those used here, originates from primesieve.