Use different segment sizes for small and large sieving primes
This is another optimization that I learned from primesieve. For small
primes, we continue to choose a segment size matching the L1 data cache
size; this is now configured using SMALL_SEGMENT_BYTES. But now we
introduce LARGE_SEGMENT_BYTES, which is configured based on the L2 cache
size (but is not necessarily equal to it; primesieve notes that using
the full L2 cache is often slower than using half). Large sieving primes
are processed using this larger segment size, which gives us a
significant performance boost when sieving higher intervals (for which
there are comparatively many large sieving primes in play).

Practically speaking, to sieve a large segment, we first divide it into
a number of small segments and then mark multiples of small sieving
primes on each of those small segments serially. (Thus we require
LARGE_SEGMENT_BYTES to be a multiple of SMALL_SEGMENT_BYTES, and this is
enforced by preprocessor directives.) Then we mark multiples of large
sieving primes on the entire large segment.